requirements.txt
ccxt
pandas
python-dotenv
requests

